Auth logic

+++ src/avatar/avatar.service.ts
@@ -1,4 +1,9 @@
-import { Injectable, Logger, NotFoundException } from '@nestjs/common';
+import {
+  ForbiddenException,
+  Injectable,
+  Logger,
+  NotFoundException,
+} from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Avatar } from './entities/avatar.entity';
 import { Repository } from 'typeorm';
@@ -31,8 +36,29 @@ export class AvatarService {
     return avatar;
   }
 
-  async remove(id: number) {
-    const avatar = await this.findOne(id);
+  async remove(userId: number, id: number) {
+    const avatar = await this.repository.findOne({
+      where: { id },
+      relations: {
+        person: true,
+      },
+    });
+
+    if (!avatar) {
+      const notFoundException = new NotFoundException('Avatar not found');
+
+      this.logger.error('Unable to find avatar', notFoundException.stack);
+
+      throw notFoundException;
+    }
+
+    if (userId !== avatar.person.id) {
+      const forbiddenException = new ForbiddenException('No access');
+
+      this.logger.error('No access', forbiddenException.stack);
+
+      throw forbiddenException;
+    }
 
     fs.unlink(FILES_PATH + avatar.filename, (err) => {
       if (err) {
